DowntoDebate Code (Zipped due to large size)
This is a project I've been working around since September. You can see it being run at: down-to-debate.com. I added multiple comments to clarify what part corresponds to what, so if you wish to make something similar you can easily refer to my code. This isn't the final version, as we are still in development.